Make community_login() look users up in the old system as well
Instead of requiring all users to log in to the main website once,
this allows them to keep using their old password on any aux sites
that uses the community login system.

We might want to consider actually migrating the accounts to the new
system at some point in the future, but for now, just allow logins
against the old tables, and have migration only happen when the
user logs into the main website.