Stops sending M2M updates from PgModel objects upon every single update
Prevents changes from M2M updates from appearing after every single
save to a PgModel derived class. The unfortunate side effect is that
no M2M updates will appear, but this has to do with the context in how
the diff between two objects it taking place.  An expanded note appears in
the explanation to #110

Fixes #110