Make it possible to turn off visibilty of versions in the feature matrix
Turn the feature matrix fields into their own class and add a paramter
for visible_default to it. Turning this off will hide them from view.
In the future this could be expanded to make it dynamic, but for now let's
just stick to being able to statically hide them.