Detect HTML in markdown submission forms
This will do simple html detection based on trivial regexps. If HTML is found,
it will replace it with HTML_REMOVED and also show a warning (in red color,
no less) that this is not permitted.

The actual filtering is still done server-side later, so the fact that these
regexps are trickable if you really want to doesn't matter...

Fixes #164