Organizations with >1 manager is always "new style"
Missing code branch when detecting if an event was migrated - it worked
as long as the org had a single new manager, but if it had more than
one it fell through to the same codepath.

This should fix a number of events showing as migrated even though they
were obviously much newer than that.