Fix (hopefully) unicode encodings in moderation messages
It worked perfectly fine to have unicode in text fields, but if there
were unicode characters in one of the dropdown fields, then sometimes
it would not be possible to save the entries since the moderation email
generation would crash even if they run through the same codepath.

Hopefully this fix will take care of some of the random errors that have
shown up with submissions - there might be more issues like it around the
code, however.