Add filters to feature matrix.
As the feature matrix is useful for seeing how far PostgreSQL has come, we
still want to keep older versions on display.  However, this is causing
problems displaying the newer versions on smaller screens.

This change adds a filter which only shows supported versions by default,
and allows folk to choose which versions they wish to compare.  This will
have no effect on browsers with Javascript disabled.