Support listname in the URL, not just the id
This will make it easier to link to a search form that properly
pre-populates the dropdown. This will be followed with a patch to the
archives to generate links this way.

It's not strictly speaking needed to support this as we currently
synchronize the list id between the systems, but that's something that
is likely going away in the future.