Allow contributor emails to be blank
Over the years, contributors come and go and we have many cases now where
the contributor's email address which we have is not valid.  As we really
don't want to show incorrect information on the webpage, allow contributor
email addresses to be blank for cases where we've discovered that the email
address is no longer valid.