Update roadmap reflect releases, et al
The next upcoming major release would now be 11, not 10, and it's coming
in the third quarter of 2018.

Also, add in the August planned releases to fill out a complete year.
Hopefully by having the minor release dates cover a year it'll be
clear that whenever we do a release set we remove one and add one at the
bottom, instead of randomly adding/removing them at various times.