Replace <b> and <i> tags with <strong> and <em> respectively.
The semantics of <b> and <i> tags have changed as the web has
evolved. <b> is specified to only be used as a last resort
whereas <i> can be used for other purposes (e.g. rendering icons).
Bold and italic text should be rendered using <strong> and <em>
respectively.

This patch makes this replacement in all pages except for pages
that will be deprecrated soon and in the presskits.