Append "/" to pgweb hosted URLs that currently lack one.
By default, Django expects URLs to end with a "/" but if it
accepts a request without a trailing slash, it will issue a
permanent redirect with the slash appended. While this is great,
it does mean an extra request to the server.

This patch adds a "/" to anchor tags that did not have one
already appended to it, thus helping to save time for a user
browsing .org as well as for the server.