New CSS base for the PostgreSQL website.
This also adds Bootstrap, Font Awesome to the codebase with license info.

Bootstrap and Font Awesome are CSS and font frameworks respectively
that ease modern web development.

The new CSS allows the PostgreSQL.org website to be responsive
based on browser window size as well as provide a modern look
and feel.

The redesign is built on top off the Bootstrap and Font Awesome
CSS and font frameworks respectively.

Authors: Sarah Conway <sarah.conway@crunchydata.com> and me