Apply modern pgweb styles to web-based documentation.
This patch brings the web-based documentation to match the current styles
on the main postgresql.org website. The CSS supports the generated markup
from the main PostgreSQL documentation and supports the new/old class
that are visible in the documentation (e.g. uppercase classes pre-v10).

Authors: Sarah Conway, Jonathan Katz