Break up "Awaiting Moderation" and "Approved" articles/events/etc.
This was becoming illegible and also hard to maintain based on how
we changed up the look/feel of the entire site. That way, we have
two distinct categories for both types of objects, which makes it
easier for the user to understand what is what.

Also reworded the explanation for what is going on, both based on
the structural change and because it read like a literal interpretation
of the code.