Make "Support" section on "Contact" page more prominent.
There has been a recent uptick in emails asking general questions and
support going to special purpose mailing lists. This renames the
"Technical Support" section to "Support," moves is further up the page,
and rewords it to help guide people to the appropriate places to help
get their questions answered.