Update migrations for py3
A lot of strings in the migrations needed a change from b' to ' to work
after the migration to py3, but this was forgotten in the big patch.

No actual changes included in this patch.
A lot of strings in the migrations needed a change from b' to ' to work
after the migration to py3, but this was forgotten in the big patch.