Hide Survey
Per discussion in $THREAD, it Survey is unmaintained. While it offers
some nice user interactivity, because it is unmaintained it should not
be visible on the website.

It is only commented out, so that we can re-enable in the future. The
archives will remain available, albeit obfuscated.

Thread: https://postgr.es/m/BBF56CF4-2065-45C5-B61F-D616F0CB2743%40yesql.se