Respect valign="middle" property in table headers
As part of ongoing changes to the structure of tables in the
documentation[1], the HTML property valign is being set to middle
in the table headers, but was not being respected in the pgweb CSS
rules.

This adjusts the table headers to respect the valign property, which
in testing works with both the latest documentation as well as previous
versions (and given most, if not all, table headers to date are not
multi-lined).

[1] https://git.postgresql.org/gitweb/?p=postgresql.git;a=commitdiff;h=e894c61836e4b967f7ec65358fdaed2ba86ed238