Add documentation styles for updated function & operator tables
The new documentation contains a new look & feel for how functions
and operator information are displayed, but this requires some
adjustments to the stylesheet.

This also removes from of the "!important" tags that are used
in top-level parts of the code, as we do not want those to cascade
down to lower layers of the stylesheets.