Update doc styles for catalog tables
As part of the ongoing effort to make the documentation
both easier to read visually and more friendly for the
PDF building, there have been improvements to the layout
of the reference tables.

This set of changes focuses on the catalog tables that
detail the information in the different PostgreSQL catalogs.
There is also a small adjustment for the function tables, as
one of the CSS classes was renamed in the SGML source.

Author: Tom Lane <tgl@sss.pgh.pa.us>