Remove dynamic "docs" stylesheet
This was never used. And it was probably good that it wasn't, because if
it had it would've included both two copies of bootstrap in the page,
and also the bootstrap css map (as part of the css, and not as the map).

The docs pages these days use the same base stylesheet as the rest of
the site.