Add a github action to validate syntax post-commit
While we don't use the github workflow for our development, it doesn't
hurt to get a post-commit notification when somebody forgot to validate
things pre-commit (using the hook we have). This hook also runs with
exactly the version of python/pycodestyle that we are using in
production, which might help someone.