Restructure coc templates
This creates a base templatefor the coc pages that will include the coc
itself and translations, the committee and the annual reports.

This removes the include files with _ in the name in the coc section,
which would previously both be directly accessible as pages and
part of the sitemap. The pages/ directory in the templates should really
only contain directly accessible pages.