Remove "procedure for approving news/events" from policy page
This should never have been there as it was not part of the policy, it
was instructions for what to do (missed when the page was migrated over
from the wiki). And at this point they are also incorrect due to the new
news posting handling, but it's better to remove them here that they
didn't belong in the first place.

As this does not contain any changes to the actual policy, this commit
does not bump the "last updated" date of the page.