Fix some consistency and formatting in news/events policy
Consistently talk about postgresql.org vs www.postgresql.org, and call
the team "web team" (not a mix of teams and groups, with and without
admin etc). Don't use markdown/wiki style for emphasis, use proper HTML.

As this does not contain any changes to the actual policy, this commit
does not bump the "last updated" date of the page.