Replace core team enumeration with link to the active core team
We missed updating this when the core team changed a while ago, which
indicates that it's silly to have the same thing defined in more than
one place. Thus, replace it with a link.

Reviewed by both -core and funds-group.

As this does not contain any changes to the actual policy, this commit
does not bump the "last updated" date of the page.