Fix bug in markdown preview of new objects
Whens submitting a new object, the instructions for how to do markdown
didn't show up until after the object had been saved once.

This turns out to be because the performance optimization to take an
empty string markdown and turn it into an empty string html wrote the
(empty) result to the wrong field, thereby overwriting the instructions.